refactor(orchestration): extract TaskCancellationService
Five methods, 262 body lines, moved verbatim. 2573 -> 2325 lines. Two
collaborators: WorkerSessionManager and QuestionFlowCore.
startWorkerCancellation becomes public on the service -- HumanQuestionService
calls it in Task 15 -- but gets no facade delegation; it was never public API.

The plan's port list named closeWorkerSession and logger, neither of which any
of the five touches, and omitted createId, which completeTaskCancellation uses.

Two placement invariants had to survive the move. reconcileParallelSlots opens
its own mutate, so both of its call sites must stay outside the enclosing
critical section -- kernel.mutate is non-reentrant and now throws where it once
silently deadlocked. And startWorkerCancellation is a bare unawaited statement,
not `void`-prefixed and not awaited; the golden fixtures pin the interleaving of
the detached chain it fires against the synchronous tail of the method.

cancelGroup needed rewiring the plan did not anticipate. Left calling the facade
delegation, it gained one microtask hop, which reordered the detached chain's
saveState against the `group.cancelled` log and turned the `cancelGroup cancels
its tasks` fixture red. Routing it straight at this.cancellation.* restores the
original hop depth. That fixture's interleaving rests on hop spacing, not on a
provable tick invariant -- it is the one place in the suite where facade
indirection is observable.

The service's reentrancy test now detects what its name claims. Asserting only
that requestTaskCancellation resolves proved nothing: the reconcile call sits in
a try/catch that swallows the error and logs it, so a nested-mutate throw is
invisible in the return value. The test now asserts the failure log never fires.
Verified both directions -- wrapping the reconcile call in a kernel.mutate turns
that assertion, and nothing else in the file, red.
